+++ tck/src/main/java/org/eclipse/sparkplug/tck/test/Monitor.java
@@ -13,6 +13,7 @@
 
 package org.eclipse.sparkplug.tck.test;
 
+import static org.eclipse.sparkplug.tck.test.common.Constants.FAIL;
 import static org.eclipse.sparkplug.tck.test.common.Constants.NOT_EXECUTED;
 import static org.eclipse.sparkplug.tck.test.common.Constants.TOPIC_PATH_DBIRTH;
 import static org.eclipse.sparkplug.tck.test.common.Constants.TOPIC_PATH_DDATA;
@@ -26,7 +27,13 @@ import static org.eclipse.sparkplug.tck.test.common.Requirements.CASE_SENSITIVIT
 import static org.eclipse.sparkplug.tck.test.common.Requirements.HOST_TOPIC_PHID_DEATH_PAYLOAD_BDSEQ;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_CASE_SENSITIVITY_SPARKPLUG_IDS;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_HOST_TOPIC_PHID_DEATH_PAYLOAD_BDSEQ;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_DEVICE_ID_CHARS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_DEVICE_ID_STRING;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_EDGE_NODE_ID_CHARS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_EDGE_NODE_ID_STRING;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_EDGE_NODE_ID_UNIQUENESS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_GROUP_ID_CHARS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_INTRO_GROUP_ID_STRING;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_MESSAGE_FLOW_DEVICE_BIRTH_PUBLISH_DBIRTH_PAYLOAD_SEQ;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_MESSAGE_FLOW_EDGE_NODE_BIRTH_PUBLISH_WILL_MESSAGE_PAYLOAD_BDSEQ;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_OPERATIONAL_BEHAVIOR_DATA_PUBLISH_DBIRTH;
@@ -53,10 +60,22 @@ import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPICS_DBIRT
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPICS_NBIRTH_BDSEQ_INCREMENT;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPICS_NBIRTH_METRIC_REQS;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPICS_NBIRTH_TEMPLATES;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_A;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_DUPLICATE_DEVICE_ID_ACROSS_EDGE_NODE;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_UNIQUE_DEVICE_ID;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_UNIQUE_EDGE_NODE_DESCRIPTOR;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.ID_TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_DEVICE_ID_STRING;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_EDGE_NODE_ID_CHARS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_EDGE_NODE_ID_STRING;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_EDGE_NODE_ID_UNIQUENESS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_GROUP_ID_CHARS;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.INTRO_GROUP_ID_STRING;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.MESSAGE_FLOW_DEVICE_BIRTH_PUBLISH_DBIRTH_PAYLOAD_SEQ;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.MESSAGE_FLOW_EDGE_NODE_BIRTH_PUBLISH_WILL_MESSAGE_PAYLOAD_BDSEQ;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.OPERATIONAL_BEHAVIOR_DATA_PUBLISH_DBIRTH;
@@ -83,9 +102,16 @@ import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPICS_DBIRTH_M
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPICS_NBIRTH_BDSEQ_INCREMENT;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPICS_NBIRTH_METRIC_REQS;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPICS_NBIRTH_TEMPLATES;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_A;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_DUPLICATE_DEVICE_ID_ACROSS_EDGE_NODE;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_UNIQUE_DEVICE_ID;
 import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_UNIQUE_EDGE_NODE_DESCRIPTOR;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID;
+import static org.eclipse.sparkplug.tck.test.common.Requirements.TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID;
 import static org.eclipse.sparkplug.tck.test.common.Utils.checkUTC;
 import static org.eclipse.sparkplug.tck.test.common.Utils.getNextSeq;
 import static org.eclipse.sparkplug.tck.test.common.Utils.getSparkplugPayload;
@@ -155,7 +181,13 @@ public class Monitor extends TCKTest implements ClientLifecycleEventListener {
 			ID_OPERATIONAL_BEHAVIOR_DATA_PUBLISH_DBIRTH_ORDER, ID_OPERATIONAL_BEHAVIOR_DATA_PUBLISH_NBIRTH_CHANGE,
 			ID_OPERATIONAL_BEHAVIOR_DATA_PUBLISH_DBIRTH_CHANGE, ID_PRINCIPLES_RBE_RECOMMENDED,
 			ID_PAYLOADS_STATE_BIRTH_PAYLOAD, ID_OPERATIONAL_BEHAVIOR_HOST_APPLICATION_HOST_ID,
-			ID_CASE_SENSITIVITY_SPARKPLUG_IDS, ID_PAYLOADS_SEQUENCE_NUM_INCREMENTING, ID_PAYLOADS_TIMESTAMP_IN_UTC };
+			ID_CASE_SENSITIVITY_SPARKPLUG_IDS, ID_PAYLOADS_SEQUENCE_NUM_INCREMENTING, ID_PAYLOADS_TIMESTAMP_IN_UTC,
+			ID_INTRO_GROUP_ID_STRING, ID_INTRO_GROUP_ID_CHARS, ID_INTRO_EDGE_NODE_ID_STRING,
+			ID_INTRO_EDGE_NODE_ID_CHARS, ID_INTRO_DEVICE_ID_STRING, ID_INTRO_DEVICE_ID_CHARS, ID_TOPIC_STRUCTURE,
+			ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES,
+			ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES,
+			ID_TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID, ID_TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID,
+			ID_TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID };
 
 	// edge_node_id to clientid
 	private HashMap<String, String> edge_nodes = new HashMap<>();
@@ -188,7 +220,7 @@ public class Monitor extends TCKTest implements ClientLifecycleEventListener {
 	HashMap<String, String> lowerGroupIds = new HashMap<String, String>();
 	HashMap<String, String> lowerEdgeIds = new HashMap<String, String>();
 	HashMap<String, String> lowerDeviceIds = new HashMap<String, String>();
-	
+
 	private Results results = null;
 
 	public Monitor(Results results) {
@@ -413,14 +445,23 @@ public class Monitor extends TCKTest implements ClientLifecycleEventListener {
 	@SpecAssertion(
 			section = Sections.PAYLOADS_B_PAYLOAD,
 			id = ID_PAYLOADS_TIMESTAMP_IN_UTC)
+	@SpecAssertion(
+			section = Sections.TOPICS_NAMESPACE_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_A)
 	@Override
 	public void publish(String clientId, PublishPacket packet) {
 
 		String topic = packet.getTopic();
-		if (topic.startsWith(NAMESPACE)) {
+
+		if (topic.startsWith("spAv1.0/")) {
+			log("Warning - non-standard Sparkplug A message received");
+			testResult(ID_TOPIC_STRUCTURE_NAMESPACE_A, setResult(false, TOPIC_STRUCTURE_NAMESPACE_A));
+		} else if (topic.startsWith(NAMESPACE)) {
 			String[] topicParts = topic.split("/");
 			// topic is spBv1.0/group_id/message_type/edge_node_id/[device_id]"
 			// or spBv1.0/STATE/hostid
+
+			checkTopic(topicParts);
 
 			if (topicParts.length > 5 || topicParts.length < 3) {
 				return;
@@ -468,6 +509,153 @@ public class Monitor extends TCKTest implements ClientLifecycleEventListener {
 				handleDDATA(group_id, edge_node_id, device_id, payload);
 			}
 		}
+	}
+
+	private void testResult(String id, String state) {
+		// Don't override a failing test fail
+		if (testResults.get(id) != null && !((String) testResults.get(id)).startsWith(FAIL)) {
+			testResults.put(id, state);
+		}
+	}
+
+	@SpecAssertion(
+			section = Sections.TOPICS_SPARKPLUG_TOPIC_NAMESPACE_ELEMENTS,
+			id = ID_TOPIC_STRUCTURE)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_GROUP_ID_STRING)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_GROUP_ID_CHARS)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_EDGE_NODE_ID_STRING)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_EDGE_NODE_ID_CHARS)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_DEVICE_ID_STRING)
+	@SpecAssertion(
+			section = Sections.INTRODUCTION_SPARKPLUG_IDS,
+			id = ID_INTRO_DEVICE_ID_CHARS)
+	@SpecAssertion(
+			section = Sections.TOPICS_GROUP_ID_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID)
+	@SpecAssertion(
+			section = Sections.TOPICS_EDGE_NODE_ID_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID)
+	@SpecAssertion(
+			section = Sections.TOPICS_DEVICE_ID_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID)
+	@SpecAssertion(
+			section = Sections.TOPICS_DEVICE_ID_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES)
+	@SpecAssertion(
+			section = Sections.TOPICS_DEVICE_ID_ELEMENT,
+			id = ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES)
+	public void checkTopic(String[] elements) {
+		Boolean result = false;
+		if (elements[0].equals(TOPIC_ROOT_SP_BV_1_0) && elements[1].equals(TOPIC_PATH_STATE)) {
+			if (elements.length == 3) {
+				result = true;
+			}
+			testResult(ID_TOPIC_STRUCTURE, setResult(result, TOPIC_STRUCTURE));
+		} else {
+			if (elements.length < 4) {
+				testResult(ID_TOPIC_STRUCTURE, setResult(false, "(too few topic elements)"));
+			} else {
+				String namespace = elements[0];
+				String group_id = elements[1];
+				String message_type = elements[2];
+				String edge_node_id = elements[3];
+				String device_id = null;
+				if (elements.length >= 5) {
+					device_id = elements[4];
+				}
+
+				if (message_type.equals("DBIRTH") || message_type.equals("DDEATH") || message_type.equals("DDATA")
+						|| message_type.equals("DCMD")) {
+
+					testResult(ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES, setResult(
+							elements.length == 5, TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_ASSOCIATED_MESSAGE_TYPES));
+					result = (elements.length == 5) ? true : false;
+				}
+
+				if (message_type.equals("NBIRTH") || message_type.equals("NDEATH") || message_type.equals("NDATA")
+						|| message_type.equals("NCMD")) {
+
+					testResult(ID_TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES, setResult(
+							elements.length == 4, TOPIC_STRUCTURE_NAMESPACE_DEVICE_ID_NON_ASSOCIATED_MESSAGE_TYPES));
+					result = (elements.length == 4) ? true : false;
+				}
+				testResult(ID_TOPIC_STRUCTURE, setResult(result, TOPIC_STRUCTURE));
+
+				result = true;
+				if (!checkUTF8String(group_id)) {
+					result = false;
+					testResult(ID_INTRO_GROUP_ID_STRING, setResult(false, INTRO_GROUP_ID_STRING));
+				}
+
+				if (!checkMQTTChars(group_id)) {
+					result = false;
+					testResult(ID_INTRO_GROUP_ID_CHARS, setResult(false, INTRO_GROUP_ID_CHARS));
+				}
+				testResult(ID_TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID,
+						setResult(result, TOPIC_STRUCTURE_NAMESPACE_VALID_GROUP_ID));
+
+				result = true;
+				if (!checkUTF8String(edge_node_id)) {
+					result = false;
+					testResult(ID_INTRO_EDGE_NODE_ID_STRING, setResult(false, INTRO_EDGE_NODE_ID_STRING));
+				}
+
+				if (!checkMQTTChars(edge_node_id)) {
+					result = false;
+					testResult(ID_INTRO_EDGE_NODE_ID_CHARS, setResult(false, INTRO_EDGE_NODE_ID_CHARS));
+				}
+				testResult(ID_TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID,
+						setResult(result, TOPIC_STRUCTURE_NAMESPACE_VALID_EDGE_NODE_ID));
+
+				if (device_id != null) {
+					result = true;
+					if (!checkUTF8String(device_id)) {
+						result = false;
+						testResult(ID_INTRO_DEVICE_ID_STRING, setResult(false, INTRO_DEVICE_ID_STRING));
+					}
+
+					if (!checkMQTTChars(device_id)) {
+						result = false;
+						testResult(ID_INTRO_DEVICE_ID_CHARS, setResult(false, INTRO_DEVICE_ID_STRING));
+					}
+					testResult(ID_TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID,
+							setResult(result, TOPIC_STRUCTURE_NAMESPACE_VALID_DEVICE_ID));
+				}
+			}
+		}
+	}
+
+	private boolean checkUTF8String(String inString) {
+		// MUST be a valid UTF-8 string
+
+		byte[] bytes = inString.getBytes(StandardCharsets.UTF_8);
+
+		String utf8EncodedString = new String(bytes, StandardCharsets.UTF_8);
+
+		boolean rc = false;
+		if (inString.equals(utf8EncodedString)) {
+			rc = true;
+		}
+		return rc;
+	}
+
+	private boolean checkMQTTChars(String inString) {
+		// MUST not use reserved characters of + (plus), / (forward slash), and # (number sign).
+		boolean rc = true;
+		if ((inString.indexOf('+') != -1) || (inString.indexOf('/') != -1) || (inString.indexOf('#') != -1)) {
+			rc = false;
+		}
+		return rc;
 	}
 
 	@SpecAssertion(
